% Generated by roxygen2 (4.0.2): do not edit by hand
\name{call_from}
\alias{call_from}
\title{Call functions and get objects from external libraries}
\usage{
call_from(pkg, name, ..., path = ".")
}
\description{
These functions allow the user to call functions and get objects
from any package installed globally while working in local mode,
without installing or even explicitly loading that package.
Use the \code{_internal} versions to access non-exported elements.
}
\details{
If the global libraries are present in \code{.libPath()} these function
are equivalent to \code{::} and \code{:::}. If not, they temporarily
switch to \code{on(soft = TRUE)}, then do the same thing.
}

