% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean_default_library}
\alias{clean_default_library}
\title{Clean the default library of user-installed packages}
\usage{
clean_default_library()
}
\description{
This function moves all the non-system packages (those with \code{"Priority" = NA}) from the
default library \code{\link{.Library}} to the user library specified in the environment
variable \code{\link{R_LIBS}}. If a package is present in both libraries, it is just removed
from the default library.
}

