% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_pacman_subdir}
\alias{create_pacman_subdir}
\title{Create a packrat sub-directory}
\usage{
create_pacman_subdir(path = ".")
}
\arguments{
\item{path}{Root directory of project.}
}
\description{
This function create a packrat sub-directory for a project.
}

