% Generated by roxygen2 (4.0.2): do not edit by hand
\name{globalise}
\alias{globalise}
\title{Call a global function}
\usage{
globalise(pkg::fun)(...)
}
\description{
Call a function using the global libraries (in addition to the local ones) when in pacman
mode, without polluting the local namespace.
}

