% Generated by roxygen2 (4.0.2): do not edit by hand
\name{init}
\alias{init}
\title{Initialise a pacman project
This function creates a pacman project in a given folder}
\usage{
init(path = ".", enter = TRUE)
}
\arguments{
\item{path}{Root directory of project}

\item{enter}{Boolean: enter pacman mode after init?}
}
\description{
Initialise a pacman project
This function creates a pacman project in a given folder
}

