% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inspect_capsule}
\alias{inspect_capsule}
\title{Inspect the capsule}
\usage{
inspect_capsule()
}
\description{
Inspect the capsule
}
\details{
Taking inspiration from the source code in \code{https://github.com/wch/R6},
  a capsule (internal environment) is used to hold meta data regarding the
  \code{\link{on}()} and \code{\link{off}()} modes. \code{inspect_capsule()} allows
  the user to inspect the contents of the capsule. It is intended for debugging only.
}

