% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lib_trees}
\alias{lib_trees}
\alias{off}
\alias{on}
\alias{pacman_mode}
\title{Set library trees}
\usage{
on(path = ".", keep_attached = NULL, keep_namespace = NULL)

off()

pacman_mode()
}
\arguments{
\item{path}{Root directory of project.}
}
\description{
Set library trees
}

