% Generated by roxygen2 (4.0.2): do not edit by hand
\name{on_off}
\alias{off}
\alias{on}
\alias{on_off}
\title{Turn pacman mode on or off}
\usage{
on(path = ".")

off()
}
\arguments{
\item{path}{Path to root of pacman project.}
}
\description{
Turn pacman mode on or off
}
\details{
These functions are used to turn pacman mode on and off. In addition to changing
  the \code{\link{.libPaths}()} for packages, they will restore the
  \code{\link{loadedNamespaces}()} and \code{\link{attach}ed} packages using the respective
  library trees.
}
\seealso{
\code{\link{pacman_mode}()} which allows the user to see which mode they are
  currently in.
}

