% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p_extract.R
\name{p_extract}
\alias{p_extract}
\title{Convert String With Commas Into Elements}
\usage{
p_extract(x, use.names = TRUE)
}
\arguments{
\item{x}{A character string of packages separated by commas; for example
the strings returned from \code{\link[pacman]{p_information}}.}

\item{use.names}{logical.  If \code{TRUE} package names, including version
number, are used.}
}
\value{
Returns a character vector of packages.
}
\description{
\code{p_extract} is designed to be used in conjunction with
\code{\link[pacman]{p_information}} to convert a single comma separated
string into a vector of package names.
}
\examples{
\dontrun{
p_extract(p_info(ggplot2, "Depends"))
p_extract(p_info(ggplot2, "Imports"))
lapply(p_info(ggplot2, "Imports", "Depends", "Suggests"), p_extract)
}
}
\seealso{
\code{\link[pacman]{p_information}}
}

