\name{p_get}
\alias{p_get}
\alias{p_install}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Installs and loads a package provided the package is a CRAN package
}
\description{
Installs a package provided the package is a CRAN package
}
\usage{
p_get(package)
p_install(package)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{package}{
name of package
}
}
\seealso{
\code{\link[utils]{install.packages}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ install}
\keyword{ package}% __ONLY ONE__ keyword per line
