% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_install_gh.R
\name{p_install_gh}
\alias{p_install_gh}
\title{Installs & Loads GitHub Packages}
\usage{
p_install_gh(package, dependencies = TRUE, ...)
}
\arguments{
\item{package}{Repository address(es) in the format 
\code{username/repo[/subdir][@ref|#pull]}.  
Note that this must be a character string.}

\item{dependencies}{logical.  If \code{TRUE} necessary dependencies will be 
installed as well.}

\item{\ldots}{Additional parameters to pass to \code{\link[remotes]{install_github}}.}
}
\description{
Installs a GitHub package.  A wrapper for \code{\link[remotes]{install_github}}
which is the same as \code{\link[devtools]{install_github}}.
}
\examples{
\dontrun{
p_install_gh("trinker/pacman")

## Package doesn't exist
p_install_gh("trinker/pacmanAwesomer")
}
}
\seealso{
\code{\link[remotes]{install_github}}
}
\keyword{github}
\keyword{install}
