% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_install_version.R
\name{p_install_version}
\alias{p_install_version}
\title{Install Minimal Package Version}
\usage{
p_install_version(package, version)
}
\arguments{
\item{package}{\code{character} vector of the name of the package(s) you want 
to install a particular minimal version of.}

\item{version}{Corresponding \code{character} vector of the minimal 
package version(s).}
}
\description{
Install minimal package version(s).
}
\examples{
\dontrun{
p_install_version(
    c("pacman", "testthat"), 
    c("0.2.0", "0.9.1")
)
}
}
\keyword{version}
