% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_install_version_gh.R
\name{p_install_version_gh}
\alias{p_install_version_gh}
\title{Install Minimal GitHub Package Version}
\usage{
p_install_version_gh(package, version, dependencies = TRUE)
}
\arguments{
\item{package}{\code{character} vector of the repository address(es) of the 
package(s) you want to install a particular minimal version of.  Repository 
address(es) in the format \code{username/repo[/subdir][@ref|#pull]}.}

\item{version}{Corresponding \code{character} vector of the minimal 
package version(s).}

\item{dependencies}{logical.  If \code{TRUE} necessary dependencies will be 
installed as well.}
}
\description{
Install minimal GitHub package version(s).
}
\examples{
p_install_version_gh(
    c("trinker/pacman", "hadley/testthat"), 
    c("0.2.0", "0.9.1")
)
}
\keyword{version}
