% Generated by roxygen2 (4.0.1): do not edit by hand
\name{p_isglobal}
\alias{p_isglobal}
\title{Test If Environment is Global}
\usage{
p_isglobal(n = 1)
}
\arguments{
\item{n}{The number of generations to go back.  If used as a function
argument n should be set to 2.}
}
\value{
A logical response.
}
\description{
A logical test to determine if the current environment is the global
environment.
}
\examples{
p_isglobal()
lapply(1:3, function(i) p_isglobal())
FUN <- function() p_isglobal(); FUN()

FUN2 <- function(x = p_isglobal(2)) x
FUN2()
FUN3 <- function() FUN2(); FUN3()
}
\author{
Simon O'Hanlon and Tyler Rinker <tyler.rinker@gmail.com>
}
\references{
\url{http://stackoverflow.com/questions/18637656/detect-if-environment-is-global-enviroment}
}
\seealso{
\code{\link[base]{globalenv}},
\code{\link[base]{parent.frame}}
}

