% Generated by roxygen2 (4.0.1): do not edit by hand
\name{p_isinstalled}
\alias{p_isinstalled}
\title{Checks if Package is Installed}
\usage{
p_isinstalled(...)
}
\arguments{
\item{...}{Name of package you want to check.  This can be quoted or unquoted.}
}
\description{
Check if package is installed locally.
}
\examples{
\dontrun{
p_installed(pacman)
p_installed(fakePackage)
}
}
\keyword{installed}
\keyword{package}

