\name{p_isloaded}
\alias{p_isloaded}
\title{Check if Package(s) is Loaded}
\usage{
p_isloaded(...)
}
\arguments{
  \item{\ldots}{Name(s) of package(s).}
}
\description{
Check if package(s) is loaded.
}
\examples{
\dontrun{
p_isloaded(ggplot2)
p_isloaded(ggplot2, dfs, pacman)
}
}
\seealso{
\code{\link[base]{.packages}}
}
\keyword{delete}
\keyword{package}

