\name{p_news}
\alias{p_news}
\title{View news for R or add-on packages}
\usage{
  p_news(package = NULL)
}
\arguments{
  \item{package}{Name of package (default is to see news
  for R)}
}
\description{
  Find out news on a package
}
\examples{
p_news()
p_news(lattice)
}
\seealso{
  \code{\link[utils]{news}}
}
\keyword{news}
\keyword{packages}

