% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_old.R
\name{p_old}
\alias{p_old}
\title{Compare Installed Packages with CRAN-like Repositories}
\usage{
p_old()
}
\value{
Retuns a \code{\link[base]{data.frame}} with info regarding out of data packages.
}
\description{
Indicates packages which have a (suitable) later version on the repositories
}
\examples{
\dontrun{
p_old()
}
}
\seealso{
\code{\link[utils]{old.packages}}
}
\keyword{old}
\keyword{outdated}

