\name{p_path}
\alias{p_path}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
See path to library of add on package
}
\description{
See path to library of add on package
}
\usage{
p_path()
}
%- maybe also 'usage' for other objects documented here.
\seealso{
\code{\link[base]{.libPaths}}
}
\examples{
p_path()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ package }
\keyword{ library }
\keyword{ location }
\keyword{ path } 
