\name{p_set_cranrepo}
\alias{p_set_cranrepo}
\title{Check if a repo is set}
\usage{
p_set_cranrepo(default_repo = "http://cran.r-project.org/")
}
\arguments{
  \item{default_repo}{The defaut package repository.

  TODO: make default_repo a package option?}
}
\description{
Check if a repo is already set and if not choose an
appropriate repo.
}

