% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pacman_project}
\alias{create_pacman_subdir}
\alias{init}
\alias{pacman_project}
\title{Initialise a pacman project}
\usage{
init(path = ".", enter = TRUE)

create_pacman_subdir(path = ".")
}
\arguments{
\item{path}{Root directory of project}

\item{enter}{Boolean: enter pacman mode after init?}
}
\description{
Initialise a pacman project
}

