% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.p_delete}
\alias{print.p_delete}
\title{Prints a p_delete Object}
\usage{
\method{print}{p_delete}(x, quiet = NULL, ...)
}
\arguments{
\item{x}{The p_delete object.}

\item{quiet}{logical. If \code{TRUE} no messages confirming package deletions
are printed.}

\item{\ldots}{ignored}
}
\description{
Prints a p_delete object.
}

