% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_version.R
\name{print.p_version_diff}
\alias{print.p_version_diff}
\title{Prints a p_version_diff Object}
\usage{
\method{print}{p_version_diff}(x, ...)
}
\arguments{
\item{x}{The p_version_diff object.}

\item{\ldots}{ignored}
}
\description{
Prints a p_version_diff object.
}

