% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_search_any.R
\name{print.search_any}
\alias{print.search_any}
\title{Prints a search_any Object}
\usage{
\method{print}{search_any}(x, ...)
}
\arguments{
\item{x}{The search_any object.}

\item{\ldots}{ignored}
}
\description{
Prints a search_any object.
}

