% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.wide_table}
\alias{print.wide_table}
\title{Prints a wide_table Object}
\usage{
\method{print}{wide_table}(x, right = FALSE, ...)
}
\arguments{
\item{x}{The wide_table object.}

\item{right}{logical.  If \code{FALSE} stings will be left-alinged.}

\item{\ldots}{ignored}
}
\description{
Prints a wide_table object.
}

