% Generated by roxygen2 (4.0.2): do not edit by hand
\name{restore_from_src}
\alias{restore_from_src}
\title{Restore from source}
\usage{
restore_from_src(path = ".")
}
\arguments{
\item{path}{Root directory of the project.}
}
\description{
Restore the private library using source files obtained using \code{\link{get_source}()}.
}

