% Generated by roxygen2 (4.0.1): do not edit by hand
\name{p_author}
\alias{p_author}
\title{Package Author}
\usage{
p_author(package = "base")
}
\arguments{
\item{package}{Name of the package you want the author of.}
}
\description{
Returns the author of a package.
}
\examples{
p_author(pacman)
p_author()
}
\seealso{
\code{\link[utils]{packageDescription}}
}
\keyword{author}
\keyword{package}

