% Generated by roxygen2 (4.0.2): do not edit by hand
\name{p_base}
\alias{p_base}
\title{Base Install Packages}
\usage{
p_base(base.only = FALSE, open = FALSE, basemarker = "***")
}
\arguments{
\item{base.only}{logical.  If \code{TRUE} a character vector of only base
install packages is returned.}

\item{open}{logical. If \code{TRUE} opens the directory of the base install
packages.}

\item{basemarker}{Character string.  The string to append to mark which
packages are part of the default packages.}
}
\description{
List all the packages and mark those in a base install.
}
\note{
Packages that are installed when R starts are marked with an asterisk(*).
}
\examples{
\dontrun{
p_base()
p_base(TRUE)
}
}
\seealso{
\code{\link[base]{getOption}}
}
\keyword{base}
\keyword{package}

