\name{p_citation}
\alias{p_citation}
\alias{p_cite}
\title{Generate citation for a package}
\usage{
  p_citation(package = "r")

  p_cite(package = "r")
}
\arguments{
  \item{package}{Name of the package you want a citation
  for}
}
\description{
  Generate citation for a package
}
\examples{
p_citation()
p_cite(pacman)
p_citation(pacman)
}
\seealso{
  \code{\link[utils]{citation}}
}
\keyword{citation}
\keyword{cite}

