% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_cran.R
\name{p_cran}
\alias{p_cran}
\alias{p_iscran}
\title{CRAN Packages}
\usage{
p_cran(menu = FALSE)

p_iscran(package)
}
\arguments{
\item{menu}{logical.  If \code{TRUE} allows user to select the package and 
return that package name.}

\item{package}{Name of package.}
}
\description{
\code{p_cran} - Generate a vector of all available packages.

\code{p_iscran} - Logical check if a package is available on CRAN.
}
\examples{
\dontrun{
p_cran()
p_cran(TRUE)
p_iscran(pacman)
}
}
\seealso{
\code{\link[utils]{available.packages}}
}
\keyword{CRAN}
\keyword{package}
