% Generated by roxygen2 (4.0.0): do not edit by hand
\name{p_data}
\alias{p_data}
\title{Package Data Sets}
\usage{
p_data(package = "datasets", static = FALSE)
}
\arguments{
\item{package}{name of package (default is the base install datasets
package).}

\item{static}{logical.  If \code{TRUE} a static text document is returned
(e.g. \code{data("datasets")}).}
}
\value{
Returns the data sets of a package as dataframe
(\code{static = FALSE}) or as a static text file (\code{static = TRUE}).
}
\description{
Generate a script of all data sets contained in package.
}
\examples{
p_data()
p_data(lattice)
\dontrun{p_data(static=TRUE)}
}
\seealso{
\code{\link[utils]{data}}
}
\keyword{data}
\keyword{package}

