% Generated by roxygen2 (4.0.0): do not edit by hand
\name{p_delete}
\alias{p_del}
\alias{p_delete}
\title{Permanently Remove Package Removal(s) From Library}
\usage{
p_delete(..., char, character.only = FALSE)

p_del(..., char, character.only = FALSE)
}
\arguments{
\item{\ldots}{name(s) of package(s).}
}
\description{
Remove package(s) from the library permanently.
}
\section{Warning}{

Using this function will remove the package from your
library and cannot be loaded again without reinstalling the package.
}
\examples{
\dontrun{
p_delete(pacman) # You never want to run this
}
}
\seealso{
\code{\link[utils]{remove.packages}}
}
\keyword{delete}
\keyword{package}

