% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p_delete.R
\name{p_delete}
\alias{p_del}
\alias{p_delete}
\title{Permanently Remove Package Removal(s) From Library}
\usage{
p_delete(..., char, character.only = FALSE, quiet = FALSE)

p_del(..., char, character.only = FALSE, quiet = FALSE)
}
\arguments{
\item{char}{Character vector containing packages to load.  If you are calling
\code{p_delete} from within a function (or just having difficulties calling it
using a character vector input) then pass your character vector of packages
to load to this parameter directly.}

\item{character.only}{logical. If \code{TRUE} then \code{p_load} will only
accept a single input which is a character vector containing the names of
packages to load.}

\item{quiet}{logical. Passed to \code{print.p_delete} as an attribute.  If
\code{TRUE} no messages confirming package deletions are printed.}

\item{\ldots}{name(s) of package(s).}
}
\description{
Remove package(s) from the library permanently.
}
\section{Warning}{

Using this function will remove the package from your
library and cannot be loaded again without reinstalling the package.
}
\examples{
\dontrun{
p_delete(pacman) # You never want to run this
}
}
\seealso{
\code{\link[utils]{remove.packages}}
}
\keyword{delete}
\keyword{package}

