% Generated by roxygen2 (4.0.2): do not edit by hand
\name{p_exists}
\alias{p_exists}
\title{Checks if Package is On CRAN/In Local Library}
\usage{
p_exists(package, cran = TRUE)
}
\arguments{
\item{package}{Name of package.}

\item{cran}{logical.  If \code{TRUE} checks CRAN for the package;
if \code{FALSE} checks the user's local library.}
}
\description{
Checks CRAN to determine if a package exists.
}
\examples{
\dontrun{
p_exists(pacman)
p_exists(pacman, FALSE)
p_exists(I_dont_exist)
}
}
\keyword{exists}
\keyword{package}

