% Generated by roxygen2 (4.0.0): do not edit by hand
\name{p_help}
\alias{p_help}
\title{Package Help Manual}
\usage{
p_help(package = NULL, web = FALSE, pdf = FALSE)
}
\arguments{
\item{package}{Name of package.}

\item{web}{logical.  If \code{TRUE} grabs current help manual from the web.}

\item{pdf}{logical.  If \code{TRUE} uses a LaTeX compiler to generate a pdf.}
}
\description{
Generate an html, web or pdf of a package's help manual.
}
\section{Warning}{

To use the pdf argument you must have a pdf compiler (e.g.
\href{http://miktex.org/}{MikTex} or
\href{https://www.tug.org/texlive/}{Tex Live}) installed.
}
\examples{
\dontrun{
p_help()
p_help(pacman)
p_help(pacman, web=TRUE)
p_help(pacman, pdf=TRUE)
}
}
\references{
\url{http://r.789695.n4.nabble.com/Opening-package-manual-from-within-R-td3763938.html}
}
\seealso{
\code{\link[utils]{help}}
}
\keyword{help}
\keyword{manual}
\keyword{package}

