% Generated by roxygen2 (4.0.2): do not edit by hand
\name{p_help}
\alias{p_help}
\title{Package Help Manual}
\usage{
p_help(package = NULL, web = FALSE, build.pdf = FALSE)
}
\arguments{
\item{package}{Name of package.}

\item{web}{logical.  If \code{TRUE} grabs current pdf help manual from the
web (\code{pdf} argument is ignored).}

\item{build.pdf}{logical.  If \code{TRUE} attempts to locate the file first
and ten uses a LaTeX compiler to generate a pdf.}
}
\description{
Generate an html, web or pdf of a package's help manual.
}
\section{Warning}{

Setting \code{build.pdf = TRUE} requires the user to have a pdf compiler (e.g.,
\href{http://miktex.org/}{MikTex} or
\href{https://www.tug.org/texlive/}{Tex Live}) installed.
}
\examples{
\dontrun{
p_help()
p_help(pacman)
p_help(pacman, web=TRUE)
p_help(pacman, build.pdf=TRUE)
}
}
\references{
\url{http://r.789695.n4.nabble.com/Opening-package-manual-from-within-R-td3763938.html}
}
\seealso{
\code{\link[utils]{help}}
}
\keyword{help}
\keyword{manual}
\keyword{package}

