% Generated by roxygen2 (4.0.1): do not edit by hand
\name{p_install}
\alias{p_get}
\alias{p_install}
\title{Installs & Loads Packages}
\usage{
p_install(package, character.only = FALSE,
  path = getOption("download_path"), ...)

p_get(package, character.only = FALSE, path = getOption("download_path"),
  ...)
}
\arguments{
\item{package}{Name of package(s).}

\item{character.only}{logical.  If \code{TRUE} \code{\dots} is treated a
character string.}

\item{\ldots}{Additional parameters to pass to \code{install.packages}.}

\item{path}{The path to the directory that contains the package.  It is
convenient to set \code{download_path} in .Rprofile options to the downloads
directory.}
}
\description{
Installs a package provided the package is a CRAN package.
}
\examples{
\dontrun{p_install(pacman)}
}
\seealso{
\code{\link[utils]{install.packages}}
}
\keyword{install}
\keyword{package}

