% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p_install_gh.R
\name{p_install_gh}
\alias{p_install_gh}
\title{Installs & Loads GitHub Packages}
\usage{
p_install_gh(package, ...)
}
\arguments{
\item{package}{Repository address(es) in the format
\code{username/repo[/subdir][@ref|#pull]}.
Note that this must be a character string.}

\item{\ldots}{Additional parameters to pass to \code{\link[devtools]{install_github}}.}
}
\description{
Installs a GitHub package.  A wrapper for \code{\link[devtools]{install_github}}.
}
\examples{
\dontrun{
p_install_gh("trinker/pacman")

## Package doesn't exist
p_install_gh("trinker/pacmanAwesomer")
}
}
\seealso{
\code{\link[devtools]{install_github}}
}
\keyword{github}
\keyword{install}

