\name{p_interactive}
\alias{p_interactive}
\alias{p_inter}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Interactive package exploration
}
\description{
Interactively search through packages, looking at functions and optionally attaching the package and looking at the help page.
}
\usage{
p_interactive()
p_inter()
}
%- maybe also 'usage' for other objects documented here.
\examples{
# p_interactive()
# p_inter()

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ packages }
