\name{p_interactive}
\alias{p_inter}
\alias{p_interactive}
\title{Interactive Package Exploration}
\usage{
p_interactive()

p_inter()
}
\description{
Interactively search through packages, looking at functions
and optionally attaching the package and looking at the
help page.
}
\examples{
\dontrun{
p_interactive()
p_inter()
}
}
\keyword{package}

