% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_library.R
\name{p_library}
\alias{p_library}
\alias{p_lib}
\title{Display Library Packages}
\usage{
p_library(open = FALSE)

p_lib(open = FALSE)
}
\arguments{
\item{open}{logical. If \code{TRUE} opens the directory of the add on 
packages library.}
}
\description{
Generates a vector of all packages available to the user and optionally opens 
the user's library (this isn't necessarily where all of the available 
packages are stored).
}
\examples{
p_lib()
p_library()
\dontrun{p_lib(TRUE)}
}
\keyword{library}
\keyword{package}
