\name{p_news}
\alias{p_news}
\title{Package/R News}
\usage{
p_news(package = NULL)
}
\arguments{
  \item{package}{Name of package (default is to see news
  for R).}
}
\description{
Find out news on a package or R.
}
\examples{
p_news()
p_news(lattice)
}
\seealso{
\code{\link[utils]{news}}
}
\keyword{news}
\keyword{packages}

