% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_opendir.R
\name{p_opendir}
\alias{p_opendir}
\title{Attempts to open a directory in a file browser}
\usage{
p_opendir(dir = getwd())
}
\arguments{
\item{dir}{A character string representing the path
(either relative or absolute) to the directory to be opened.
Defaults to the working directory.}
}
\description{
Attempts to open a directory in a file browser.
Opening a directory isn't a platform independent
but it is used in more than one function so moving
this functionality to its own non-exported function
makes sense.
}
\section{Note}{
 Most likely this function will move to a different
package at some point as it's not specifically package related.
}

\examples{
\dontrun{
p_opendir() # opens working directory
p_opendir(path.expand("~")) # opens home directory
p_opendir(pacman:::p_basepath())
}
}
