% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p_path.R
\name{p_path}
\alias{p_path}
\title{Path to Library of Add-On Packages}
\usage{
p_path()
}
\description{
Path to library of add-on packages.
}
\examples{
p_path()
}
\seealso{
\code{\link[base]{.libPaths}}
}
\keyword{library}
\keyword{location}
\keyword{package}
\keyword{path}

