% Generated by roxygen2 (4.0.0): do not edit by hand
\name{p_path}
\alias{p_path}
\title{See path to library of add on package}
\usage{
p_path()
}
\description{
See path to library of add on package
}
\examples{
p_path()
}
\seealso{
\code{\link[base]{.libPaths}}
}
\keyword{library}
\keyword{location}
\keyword{package}
\keyword{path}

