% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p_set_cranrepo.R
\name{p_set_cranrepo}
\alias{p_set_cranrepo}
\title{Check if Repo is Set}
\usage{
p_set_cranrepo(default_repo = "http://cran.rstudio.com/")
}
\arguments{
\item{default_repo}{The default package repository.}
}
\description{
Check if a repo is already set and if not choose an appropriate repo.
}

