% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_version.R
\name{p_version}
\alias{p_version}
\alias{p_ver}
\alias{p_version_cran}
\alias{p_ver_cran}
\alias{p_version_diff}
\alias{p_ver_diff}
\title{Package Version}
\usage{
p_version(package = "R")

p_ver(package = "R")

p_version_cran(package = "R")

p_ver_cran(package = "R")

p_version_diff(package = "R")

p_ver_diff(package = "R")
}
\arguments{
\item{package}{Name of package (default returns R version).}
}
\description{
\code{p_version} - Determine what version a package is in your library.

\code{p_version_cran} - Determine what version a package is on CRAN.

\code{p_version_difference} - Determine version difference between a local 
package and CRAN.
}
\examples{
\dontrun{
p_ver()
p_version()
p_ver(pacman)
p_version(pacman)

p_ver_cran()
p_ver_cran(pacman)

## Compare local to CRAN version
p_ver(pacman) == p_ver_cran(pacman)
p_ver(pacman) > p_ver_cran(pacman)

p_ver_diff()
p_ver_diff(pacman)
}
}
\seealso{
\code{\link[utils]{packageDescription}}
}
\keyword{package}
\keyword{version}
