% Generated by roxygen2 (4.0.2): do not edit by hand
\name{restore_from_src}
\alias{restore_from_src}
\title{Restore the private library from the local source files}
\usage{
restore_from_src(path = ".")
}
\arguments{
\item{path}{Root directory of the project.}
}
\description{
This function is similar to \code{packrat::\link[packrat]{restore}()},
but it installs the packages from the sources in \code{pacman/src}, instead of
obtaining them remotely.
}

